git-commit/backend/requirements.txt
flask
flask-cors
gitpython